Update external libraries
Libraries in the "external" folder have been updated to the following
versions (latest as of 2024-03-04):
- GLEW v2.2.0 (https://github.com/nigels-com/glew)
- GLFW v3.4 (https://github.com/glfw/glfw)
- GLM v1.0.1 (https://github.com/g-truc/glm)
- ImGUI v1.90.4 (https://github.com/ocornut/imgui)
- lodepng d398e0f 2023-12-03 (https://github.com/lvandeve/lodepng)

For each, the source archive of the latest release has been included.
The CMakeLists.txt and some source files needed to be adjusted as well.

Note: It would be a good idea to switch to git submodules